Update wml proxy for Waston machine learning protocol version 2020-10-20 (Squashed)
This commit addresses changes mainly introduced by version 2020-9-1: `space_id` and `instance_id`. `instance_id` needs to be deleted from each request header and `space_id` need to be added in query string to each resource request. This commit also addresses some minor changes in request body(the location change of model id and the removal of version number)

This commit dose not fix the problem that all requests are raw HTTP requests instead of SDK invocation. This might be considered as a technical debt if we want to recommend WML proxy as an reference implementation.